Made two changes to demonstrate capability.
Changed the Generate object from a module to a class that inherents from the Base class like all other commands.  Gave it a flag options - that is an optiong of type :boolean.  Gave it a "call" instance method.  This makes the sub-commands that are regristered with "generate" as its prefix optional.  When generate was a module and regristered anonymously the sub-commands were required.

To demonstrate the new "call" method, got access to all the sub commands that have been regristered against the "generate" prefix.  Then called a new class method named "help" on each sub command.

The "generate test" command demonstrates how to obtain the standard Dry::CLI help text for use within the command.  This was put into a new class method for Generate::Test class named help.

Using `rescue SystemExit` around the invokation of the stardard Dry::CLI help system in order to ignore the `exit(0)` and using StringIO to get access to the output from the standard help functionality.